+++ xeo/__init__.py
@@ -1,6 +1,6 @@
 """xeo - Earth observation instruments in Python."""
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"
 __author__ = "David Montero Loaiza <dml.mont@gmail.com>"
 __all__ = [
     "Catalogue",